[codex] P1-M6-C2 foliage instancing and review fixes (#4)
* P1-M6-C2 foliage instancing

* Clean CI warning output

* Suppress lcov ignored warning output

* docs: add source comment coverage

* Implement shader variant cache

* Harden scene and editor lifecycle

* Harden runtime ownership and streaming

* Fix scripting binding test link path

* Record ownership stability audit

* Move physics shape payloads into world context

* Split scripting runtime binding state

* Own DAP server session state

* Remove legacy global service bindings

* Extract scripting deferred mutation queue

* Fix reviewed engine correctness issues

* Fix console help Werror warning

* Make foundation unit test headless

+++ scripting/include/engine/scripting/bindable_api.h
@@ -1,3 +1,5 @@
+// Declares bindable api types and APIs for the Engine Lua scripting system.
+
 #pragma once
 // Bindable API — thin accessors for Lua binding generation.
 // Annotated with // LUA_BIND: for use with tools/binding_generator.
@@ -61,11 +63,11 @@ const char *bindable_get_game_mode() noexcept;
 // LUA_BIND: set_game_state(name: string) -> bool
 bool bindable_set_game_state(const char *name) noexcept;
 
-// LUA_BIND: is_alive(entity: uint32) -> bool
-bool bindable_is_alive(std::uint32_t entity) noexcept;
+// LUA_BIND: is_alive(entity: uint64) -> bool
+bool bindable_is_alive(std::uint64_t entity) noexcept;
 
-// LUA_BIND: has_light(entity: uint32) -> bool
-bool bindable_has_light(std::uint32_t entity) noexcept;
+// LUA_BIND: has_light(entity: uint64) -> bool
+bool bindable_has_light(std::uint64_t entity) noexcept;
 
 // LUA_BIND: set_camera_fov(fov: float) -> void
 void bindable_set_camera_fov(float fov) noexcept;
